Function read_write_set::dynamic_analysis::concretize

pub fn concretize(
    accesses: AccessPathTrie<Access>, 
    signers: &[AccountAddress], 
    actuals: &[Vec<u8>], 
    type_actuals: &[TypeTag], 
    fun_env: &FunctionEnv<'_>, 
    blockchain_view: &dyn MoveStorage
) -> Result<ConcretizedSecondaryIndexes>

Bind all formals and type variables in accesses using signers, actuals, and type_actuals. In addition, concretize all secondary indexes in accesses against the state in blockchain_view.