Struct read_write_set::dynamic_analysis::ConcretizedFormals

pub struct ConcretizedFormals(_);

A read/write set state with no unbound formals or type variables

Implementations

impl ConcretizedFormals

pub fn get_keys_(&self, is_write: bool, env: &GlobalEnv) -> Vec<ResourceKey>

Return the ResourceKey’s that may be accessed by self. If is_write is true, return the keys that may be written; otherwise, return the keys that may be read. For example: if self is 0x7/0x1::AModule::AResource/f/g -> ReadWrite, this will return 0x7/0x1::AModule.

pub fn get_keys_written(&self, env: &GlobalEnv) -> Vec<ResourceKey>

Return the ResourceKey’s written by self.

pub fn get_keys_read(&self, env: &GlobalEnv) -> Vec<ResourceKey>

Return the ResourceKey’s read by self.

pub fn from_args(
accesses: AccessPathTrie<Access>,
signers: &[AccountAddress],
actuals: &[Vec<u8>],
type_actuals: &[TypeTag],
fun_env: &FunctionEnv<'_>
) -> Result<ConcretizedFormals>

Construct a ConcretizedFormals from accesses by binding the formals and type variables in accesses to signers/actuals and type_actuals. For example: if accesses is Formal(0)/0x1::M::S<TypeVar(0)>/f -> Read, signers is 0xA and type_actuals is 0x2::M2::S2, this will return 0xA/0x1::M::S<0x2::M2::S@>/f -> Read